Removed multi-byte characters and Byte Order Mark
REDCap apparently added the BOM when we added the multi-byte characters by copying and pasting from a Word document.  Now that those characters have been removed, it appears that we can upload without the BOM and when we download it is still not there.  I think it's better to keep it this way for simplicity.  I (or some successor) might have a different opinion in the future.