Checkpoint of git history, used to check May and early June 2020 hours (on or before June 6)
The git command to write the history is simple.  It's just:

    git log>history.log

For some reason, copy and paste from the Sourcetree commit message shown in the log into some other application does not seem to pick up everything in the message.  I'm going to check for Sourcetree updates after this commit.