Checkpoint of git history, used to hours up to today (June 18) for early invoice.
The git command to write the history is simple.  It's just:

    git log>history.log

For some reason, copy and paste from the Sourcetree commit message shown in the log into some other application does not seem to pick up everything in the message.

This is a reported known bug in Sourcetree, but you can copy selected text in the commit message using right-click and copy.