old_records_lyme is a poorly named field that contains a file with old "Version 1" data for clients who existed at the time (late 2018?), one file for each client.
I moved that field to the new version 3 profile mainly at the request of Cindy Grigel, the Thetford Community Nurse.  It's at the very end, labeled "Old Version 1 Records", so that it will still be available when we switch to Version 3 (scheduled for July 12, 2020).

I wonder how hard it was to create/upload those files.  Perhaps I could do the same thing for some of the Version 2 data?