Work in progress to implement Kristin Barnum's form changes. Change choices for encounter_type.
Three of the new choices using the same internal codes as before.
This means that:

What was "Office hours" will be
"In person face to face (office, inpatient, MD office) ".

What was "Phone, client" will be
"Phone, client/family/caregiver".

What was "Email, client" will be
"Email, client/family/caregiver".

Meaning, if we do any reports before/after the changeover we will need
to account for the change of meaning.

The other old choices will all be hidden.